Added missing impl of service
